Refactored artifact store implementation
* dependency inversion principle: runnable in-memory store
implementation is decoupled from the core by means of interfaces
and dependency injection
* moved HTTP server handler services to separate package
* renamed fuseml package to core package
