+++ pkg/core/codeset_store.go
@@ -1,4 +1,4 @@
-package fuseml
+package core
 
 import (
 	"context"
@@ -7,13 +7,6 @@ import (
 
 	"github.com/fuseml/fuseml-core/gen/codeset"
 )
-
-// CodesetStore is an inteface to codeset stores
-type CodesetStore interface {
-	Find(ctx context.Context, project, name string) (*codeset.Codeset, error)
-	GetAll(ctx context.Context, project, label *string) ([]*codeset.Codeset, error)
-	Add(ctx context.Context, c *codeset.Codeset) (*codeset.Codeset, error)
-}
 
 // GitAdmin is an inteface to git administration clients
 type GitAdmin interface {
