+++ pkg/core/runnable_store.go
@@ -1,9 +1,10 @@
-package fuseml
+package core
 
 import (
+	"context"
 	"time"
 
-	runnable "github.com/fuseml/fuseml-core/gen/runnable"
+	"github.com/fuseml/fuseml-core/gen/runnable"
 	"github.com/google/uuid"
 )
 
@@ -12,18 +13,21 @@ type RunnableStore struct {
 	items map[uuid.UUID]*runnable.Runnable
 }
 
-var (
-	runnableStore = RunnableStore{items: make(map[uuid.UUID]*runnable.Runnable)}
-)
+// NewRunnableStore returns an in-memory runnable store instance
+func NewRunnableStore() *RunnableStore {
+	return &RunnableStore{
+		items: make(map[uuid.UUID]*runnable.Runnable),
+	}
+}
 
-// FindRunnable returns a runnable identified by id
-func (rs *RunnableStore) FindRunnable(id uuid.UUID) *runnable.Runnable {
+// Find returns a runnable identified by id
+func (rs *RunnableStore) Find(ctx context.Context, id uuid.UUID) *runnable.Runnable {
 	return rs.items[id]
 }
 
-// GetAllRunnables returns all runnables of a given type.
+// GetAll returns all runnables of a given type.
 // Type can be "all" for returning runnables of all types.
-func (rs *RunnableStore) GetAllRunnables(kind string) (result []*runnable.Runnable) {
+func (rs *RunnableStore) GetAll(ctx context.Context, kind string) (result []*runnable.Runnable) {
 	result = make([]*runnable.Runnable, 0, len(rs.items))
 	for _, r := range rs.items {
 		if kind == "all" || r.Kind == kind {
@@ -33,8 +37,8 @@ func (rs *RunnableStore) GetAllRunnables(kind string) (result []*runnable.Runnab
 	return
 }
 
-// AddRunnable adds a new runnable, based on the Runnable structure provided as argument
-func (rs *RunnableStore) AddRunnable(r *runnable.Runnable) (*runnable.Runnable, error) {
+// Add adds a new runnable, based on the Runnable structure provided as argument
+func (rs *RunnableStore) Add(ctx context.Context, r *runnable.Runnable) (*runnable.Runnable, error) {
 	id, err := uuid.NewUUID()
 	if err != nil {
 		return nil, err
