+++ pkg/svc/runnable.go
@@ -1,4 +1,4 @@
-package fuseml
+package svc
 
 import (
 	"context"
@@ -7,18 +7,20 @@ import (
 
 	"github.com/google/uuid"
 
-	runnable "github.com/fuseml/fuseml-core/gen/runnable"
+	"github.com/fuseml/fuseml-core/gen/runnable"
+	"github.com/fuseml/fuseml-core/pkg/domain"
 )
 
 // runnable service example implementation.
 // The example methods log the requests and return zero values.
 type runnablesrvc struct {
 	logger *log.Logger
+	store  domain.RunnableStore
 }
 
-// NewRunnable returns the runnable service implementation.
-func NewRunnable(logger *log.Logger) runnable.Service {
-	return &runnablesrvc{logger}
+// NewRunnableService returns the runnable service implementation.
+func NewRunnableService(logger *log.Logger, store domain.RunnableStore) runnable.Service {
+	return &runnablesrvc{logger, store}
 }
 
 // Retrieve information about runnables registered in FuseML.
@@ -29,23 +31,23 @@ func (s *runnablesrvc) List(ctx context.Context, p *runnable.ListPayload) (res [
 		kind = *p.Kind
 	}
 
-	return runnableStore.GetAllRunnables(kind), nil
+	return s.store.GetAll(ctx, kind), nil
 }
 
 // Register a runnable with the FuseML runnable runnableStore.
 func (s *runnablesrvc) Register(ctx context.Context, p *runnable.Runnable) (res *runnable.Runnable, err error) {
 	s.logger.Print("runnable.register")
-	return runnableStore.AddRunnable(p)
+	return s.store.Add(ctx, p)
 }
 
-// Retrieve an Runnable from FuseML.
+// Retrieve a Runnable from FuseML.
 func (s *runnablesrvc) Get(ctx context.Context, p *runnable.GetPayload) (res *runnable.Runnable, err error) {
 	s.logger.Print("runnable.get")
 	id, err := uuid.Parse(p.RunnableNameOrID)
 	if err != nil {
 		return nil, runnable.MakeBadRequest(err)
 	}
-	r := runnableStore.FindRunnable(id)
+	r := s.store.Find(ctx, id)
 	if r == nil {
 		return nil, runnable.MakeNotFound(errors.New("could not find a runnable with the specified ID"))
 	}
